+++ packages/veui-theme-one/components/Table.js
@@ -1,6 +1,11 @@
 import config from 'veui/managers/config'
+import '../icons/angle-right'
 
 config.defaults({
+  icons: {
+    expand: 'angle-right',
+    collapse: 'angle-right'
+  },
   ui: {
     slim: {
       boolean: true