+++ packages/veui/src/components/Table/_TableFoot.js
@@ -5,9 +5,10 @@ export default {
   mixins: [table],
   computed: {
     ...table.mapTableData(
-      { realColumns: 'columns' },
       'selectable',
-      'selectStatus'
+      'expandable',
+      { realColumns: 'columns' },
+      { viewColumnCount: 'columnCount' }
     )
   },
   render () {
@@ -16,12 +17,14 @@ export default {
         <tr>
           {
             this.$slots.default
-              ? <th colspan={this.columns.length + (this.table.selectable ? 1 : 0)}>{this.$slots.default}</th>
-              : (this.table.selectable ? [<th></th>] : []).concat(
-                this.columns.map(col => (
-                  <th class={col.align ? `veui-table-column-${col.align}` : null}>{col.renderFoot()}</th>
-                ))
-              )
+              ? <th colspan={this.columnCount}>{this.$slots.default}</th>
+              : (this.selectable ? [<th></th>] : [])
+                .concat(this.expandable ? [<th></th>] : [])
+                .concat(
+                  this.columns.map(col => (
+                    <th class={col.align ? `veui-table-column-${col.align}` : null}>{col.renderFoot()}</th>
+                  ))
+                )
           }
         </tr>
       </tfoot>