+++ CHANGELOG.md
@@ -45,7 +45,8 @@
 * `Field` 组件对应的数据字段名现在默认优先取 `name` prop，且可以被 `field` prop 覆盖。
 * `rule` 模板占位符由 `${...}` 变更为 `{...}`，以方便在模板字符串中进行书写。老语法仍然保持兼容。
 * `Calendar`、`DatePicker` 与 `Schedule` 的范围合并逻辑现在默认为“智能”模式，当用户从已选项开始选择范围时，将从整体已选范围中去除当前选区；当从未选项开始选择时，则将当前选区并入整体。
-* `Table` 组件形状 prop `scroll`，用来限定内容滚动区域的最大高度，设置时会使得表格的头/脚固定。
+* `Table` 组件新增 prop `scroll`，用来限定内容滚动区域的最大高度，设置时会使得表格的头/脚固定。
+* `Table` 组件新增 prop `expandable` 与 `expanded`，用来支持行展开。
 * `Link` 组件新增 prop `rel`、`target`，当 `target` 为 `_blank` 时自动为 `rel` 增加 `noopener` 值，以增强安全性。
 
 ### 🐞 问题修复